+++ lib/src/main/java/com/diffplug/spotless/generic/LicenseHeaderStep.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,10 @@
  */
 package com.diffplug.spotless.generic;
 
+import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
 import java.io.Serializable;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
@@ -32,6 +35,8 @@ import javax.annotation.Nullable;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.LineEnding;
 import com.diffplug.spotless.SerializableFileFilter;
+
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 /** Prefixes a license header before the package statement. */
 public final class LicenseHeaderStep implements Serializable {
@@ -206,5 +211,75 @@ public final class LicenseHeaderStep implements Serializable {
 				return beforeYear + yearToday + afterYear + raw.substring(contentMatcher.start());
 			}
 		}
+	}
+
+	private static final String spotlessSetLicenseHeaderYearsFromGitHistory = "spotlessSetLicenseHeaderYearsFromGitHistory";
+
+	public static final String FLAG_SET_LICENSE_HEADER_YEARS_FROM_GIT_HISTORY() {
+		return spotlessSetLicenseHeaderYearsFromGitHistory;
+	}
+
+	/** Sets copyright years on the given file by finding the oldest and most recent commits throughout git history. */
+	public String setLicenseHeaderYearsFromGitHistory(String raw, File file) throws IOException {
+		if (yearToday == null) {
+			return raw;
+		}
+		Matcher contentMatcher = delimiterPattern.matcher(raw);
+		if (!contentMatcher.find()) {
+			throw new IllegalArgumentException("Unable to find delimiter regex " + delimiterPattern);
+		}
+
+		String oldYear;
+		try {
+			oldYear = parseYear("git log --follow --find-renames=40% --diff-filter=A", file);
+		} catch (IllegalArgumentException e) {
+			// Ideally, git log would always find the commit where it was added.
+			// For some reason, that is sometimes not possible - in that case,
+			// we'll settle for just the most recent, even if it was just a modification.
+			oldYear = parseYear("git log --follow --find-renames=40% --reverse", file);
+		}
+		String newYear = parseYear("git log --max-count=1", file);
+		String yearRange;
+		if (oldYear.equals(newYear)) {
+			yearRange = oldYear;
+		} else {
+			yearRange = oldYear + yearSepOrFull + newYear;
+		}
+		return beforeYear + yearRange + afterYear + raw.substring(contentMatcher.start());
+	}
+
+	private static String parseYear(String cmd, File file) throws IOException {
+		String fullCmd = cmd + " " + file.getAbsolutePath();
+		ProcessBuilder builder = new ProcessBuilder().directory(file.getParentFile());
+		if (LineEnding.nativeIsWin()) {
+			builder.command("cmd", "/c", fullCmd);
+		} else {
+			builder.command("bash", "-c", fullCmd);
+		}
+		Process process = builder.start();
+		String output = drain(process.getInputStream());
+		String error = drain(process.getErrorStream());
+		if (!error.isEmpty()) {
+			throw new IllegalArgumentException("Error for command '" + fullCmd + "':\n" + error);
+		}
+		Matcher matcher = FIND_YEAR.matcher(output);
+		if (matcher.find()) {
+			return matcher.group(1);
+		} else {
+			throw new IllegalArgumentException("Unable to parse date from command '" + fullCmd + "':\n" + output);
+		}
+	}
+
+	private static final Pattern FIND_YEAR = Pattern.compile("Date:   .* ([0-9]{4}) ");
+
+	@SuppressFBWarnings("DM_DEFAULT_ENCODING")
+	private static String drain(InputStream stream) throws IOException {
+		ByteArrayOutputStream output = new ByteArrayOutputStream();
+		byte[] buf = new byte[1024];
+		int numRead;
+		while ((numRead = stream.read(buf)) != -1) {
+			output.write(buf, 0, numRead);
+		}
+		return new String(output.toByteArray());
 	}
 }