+++ testlib/src/test/java/com/diffplug/spotless/npm/PrettierFormatterStepTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@ package com.diffplug.spotless.npm;
 
 import java.io.File;
 import java.util.Arrays;
+import java.util.Collections;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -41,7 +42,7 @@ public class PrettierFormatterStepTest {
 
 		@Parameterized.Parameters(name = "{index}: prettier can be applied to {0}")
 		public static Iterable<String> formattingConfigFiles() {
-			return Arrays.asList("typescript", "json", "javascript-es5", "javascript-es6", "css", "scss", "markdown", "yaml");
+			return Arrays.asList("html", "typescript", "json", "javascript-es5", "javascript-es6", "css", "scss", "markdown", "yaml");
 		}
 
 		@Test
@@ -69,7 +70,7 @@ public class PrettierFormatterStepTest {
 	public static class SpecificPrettierFormatterStepTests extends NpmFormatterStepCommonTests {
 
 		@Test
-		public void parserInferenceIsWorking() throws Exception {
+		public void parserInferenceBasedOnExplicitFilepathIsWorking() throws Exception {
 			String filedir = "npm/prettier/filetypes/json/";
 
 			final String dirtyFile = filedir + "json.dirty";
@@ -84,6 +85,41 @@ public class PrettierFormatterStepTest {
 
 			try (StepHarness stepHarness = StepHarness.forStep(formatterStep)) {
 				stepHarness.testResource(dirtyFile, cleanFile);
+			}
+		}
+
+		@Test
+		public void parserInferenceBasedOnFilenameIsWorking() throws Exception {
+			String filedir = "npm/prettier/filename/";
+
+			final String dirtyFile = filedir + "dirty.json";
+			final String cleanFile = filedir + "clean.json";
+
+			final FormatterStep formatterStep = PrettierFormatterStep.create(
+					PrettierFormatterStep.defaultDevDependencies(),
+					TestProvisioner.mavenCentral(),
+					buildDir(),
+					npmExecutable(),
+					new PrettierConfig(null, Collections.emptyMap()));
+
+			try (StepHarnessWithFile stepHarness = StepHarnessWithFile.forStep(formatterStep)) {
+				stepHarness.testResource(new File("test.json"), dirtyFile, cleanFile);
+			}
+		}
+
+		@Test
+		public void verifyPrettierErrorMessageIsRelayed() throws Exception {
+			FormatterStep formatterStep = PrettierFormatterStep.create(
+					PrettierFormatterStep.defaultDevDependenciesWithPrettier("2.0.5"),
+					TestProvisioner.mavenCentral(),
+					buildDir(),
+					npmExecutable(),
+					new PrettierConfig(null, ImmutableMap.of("parser", "postcss")));
+			try (StepHarness stepHarness = StepHarness.forStep(formatterStep)) {
+				stepHarness.testException("npm/prettier/filetypes/scss/scss.dirty", exception -> {
+					exception.hasMessageContaining("HTTP 501");
+					exception.hasMessageContaining("Couldn't resolve parser \"postcss\"");
+				});
 			}
 		}
 	}